+++ src/main/java/com/parkez/image/config/S3Config.java
@@ -1,4 +1,4 @@
-package com.parkez.common.config;
+package com.parkez.image.config;
 
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.BasicAWSCredentials;
