+++ src/main/java/com/parkez/image/dto/request/ImageDeleteRequest.java
@@ -1,4 +1,4 @@
-package com.parkez.common.image.dto.request;
+package com.parkez.image.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
@@ -8,8 +8,8 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
-@Schema(description = "이미지 업로드/수정 시 요청 DTO")
-public class ImageUploadRequest {
+@Schema(description = "이미지 삭제 시 요청 DTO")
+public class ImageDeleteRequest {
 
     @NotBlank(message = "targetType은 필수 입력값입니다.")
     @Schema(description = "이미지 업로드할 도메인 입력", example = "UserProfile")
@@ -19,9 +19,8 @@ public class ImageUploadRequest {
     @Schema(description = "이미지 업로드할 도메인의 개별 Id 입력", example = "1")
     private Long targetId;
 
-
     @Builder
-    private ImageUploadRequest(String targetType, Long targetId) {
+    private ImageDeleteRequest(String targetType, Long targetId) {
         this.targetType = targetType;
         this.targetId = targetId;
     }
